Update index.htm
--------------------------------Explaing the audio changes-------------------------------------------------


In mobile browsers autoplay doesn't work for the audio and video tag due to mobile charges. The only way to start playing audio is if the user initiates it either through a button click or something like that.

Read more here: http://stackoverflow.com/a/26066844

So I made some changes that detect the user agent to see if its a mobile browser or not (see script afunction(); on line 152). If it detects that the user is using a mobile browser, it replaces the inner html with material design lite (Google's version of material design for the web), and two button, and text that states, "Prepare Your Wallet". One button (the one that uses material design) says "KKKK" which, onclick, goes to a function called WalletPrepared() (line 62). That function starts playing the audio and makes an attribute that says, onload of a random picture (line 75) to initiate WalletPrepared2() (line 109). The reason I chose, onload of a random picture, is because, I don't know how to execute another Javascript from another Javascript (I'm a newb). Anyway, what WalletPrepared2() does is, replace all of the inner html with your code. Thus, seeing lord Gaben making a bunch of deals on Steam.

But, if the user clicks the No button, on the prompt, it will goto WalletNotPrepared() (line 42). That replace the text, "Prepare your wallet" with "Um, I don't think you heard me. I said, Prepare Your Wallet!!!!!!!!!!!!!". It also removes the No button. And when the user clicks KKKK it will goto the script on line 62 and I just explained what it does above.


But if the user is on a computer, it makes a random img, onload, to go to function WalletPrepared3() (line 67). WalletPrepared3() (line 86) is the same as WalletPrepared2(). WalletPrepared3 executes WalletPrepared2 to replace the html code except on WalletPrepared3() is that it doesn't start playing, Are You Ready For A Miracle, right away.


I only used secure sources (https) so that, not only will the Steam iframe load, but users can also access "gabegaming.com". Using "https://gabegaming.com", which prevents malicious stuff being loaded on old web browsers.

Speaking of older browsers, I added a fallback option between the <source tag and <audio, which uses <embed, to start playing the audio in html4 instead of html5.

HTML4 ignores the <audio and <source tag, but cares about the stuff in the middle.
HTML5 ignores everything in the <audio tag except <source.

// Note: If the user clicks a button. The button better pointing to a JS which runs it immediately (And then that starts playing the sound if the user //is on mobile). I tried just the button pointing to a JS which replaces all of the inner html with your old version of it. But the audio still didn't work.

Just some note: If the mobile user clicks the, "KKKK" button. The button better pointing to a JS which runs it immediately (And then that starts playing the sound if the user is on mobile). I tried just the button pointing to a JS which replaces all of the inner html with your old version of it. But the audio still didn't work.


---------Explaining changes to the Steam iFrame which makes the iFrame work (well, sort of [if the user clicks some of the links while in https mode the iFrame goes blank})

SO, apparently you can attack people using iFrames.

Read more: http://javascript.info/tutorial/clickjacking

So they thought, hey, instead of somehow implementing a system that can detect css and iFrame attacks, lets just add an option so people who don't want their website displayed in an iFrame, they can add an html header (X-Frames-Options), which loads wayyyyy, before Javascript is downloaded and interpreted. Anyway, Steam put the X-Frames-Options to DENY and that means you can't put Steam in an iFrame. All XMLHttpRequests are blocked, except there is a flaw in the system. JsonP is not effected by XMLHttpRequest. So by Googling for about 3 hours (not exaggerating), I found a workaround.

Read More: http://stackoverflow.com/a/25587483/5195963

The problem with this solution was that it requires opening an iFrame. When doing that, Chrome and other browsers error out and say you need to put the allow-same-origins flag to bypass the sandbox. I put that flag but it still didn't work. I think I read somewhere that you can set the domain to something else so it would still work, but the solution wouldn't work in Firefox, so I dismissed it. After 2 days of thinking I came up with document.write, but the problem with that, after DOM loads the page you need to use document.open to access a page. Which erases the whole html pages when running document.write with whatever content. But I still tried it, what I got was Steam in full screen. After seeing if I can somehow use the document.write api to  just add code I gave up (I can't use innerHTML because for some reason it won't recognize the "algorithm". So after being stuck on that for 1 day, I just gave up and said, Shrek this, I'll just create a seperate html page which gets erased with document.write to Steam and then reference that iFrame to gabegaming.com so it'll work. Also, I tried using a slow page down service which slows the page down so DOM won't finish loading, but that just leaved an empty page.


Outcome=Everything Works as expected.

You can scroll and everything, but there is another problem, there isn't an unlimited page scroll like if you visited the webpage, but its better than a screenshot or just a picture. Main Point is, you can see the sales on Steam.


=========================Bottom line====================

I added https support, Steam iFrame, and fixed some audio issues, including html4 and on mobile.

Note: I didn't use an html4 browser but when I get a chance I'll test it.




P.S. If you look OGre the code, you can see all my comments explaining what it does

P.P.S. Also, I tested this on my computer using the horrible browser (Microsoft Edge), Chrome, and Firefox, on my phone using Chrome, and I sent my friend a link to the sample page I uploaded to github to see if the http/https fix would work. Everything works as expected. So please except this file change because it took me about 2 months.

P.P.P.S. The audio loops on both mobile and desktop.

P.P.P.P.S. I used detectmobilebrowsers.com and Stackoverflow. I don't have a tablet. But this should work on it to.


Example page I uploaded: http://rawgit.com/tvman-123/defaultnamehere.github.io/master/Gabegaming_With_HTML.html and the crappy https version because Steam wasn't met to be displayed that way

 https://rawgit.com/tvman-123/defaultnamehere.github.io/master/Gabegaming_With_HTML.html


P.P.P.P.P.S.: If you can see deleted commants on GIthub, lol, I'm sorry I said that adding https to store.steampowered will fix it. The only reason I said that as I downloaded a plugin that removed the X-Frames-Option header, but at the same time I added the s to htp and it worked. Sooooooo sad....